Add basic GraphQL functionality. Change contributors to issues.
Connected to GitHub GraphQL API to test fetching repos with a pre-written call. This design is not viable as distribution would require sharing an access token. Issues are being used instead of contributors since GitHub's GraphQL API does not yet offer the ability to fetch the number of contributors for repositories that the viewer does have push access for.